+++ RELEASE_NOTES.rst
@@ -17,6 +17,8 @@ API.
 2.2.0
 ~~~~~
 
+- API: DF estimators now support native estimators using sparse matrices as input or output, and automatically
+  convert them to/from sparse :class:`~pandas.DataFrame` objects
 - API: new property :attr:`.EstimatorDF.output_names_` to get the names of the output
   columns the estimator was fitted with
 - API: new method :attr:`.LearnerPipelineDF.preprocess` to apply the preprocessing step